+++ .gitignore
@@ -2,3 +2,5 @@ env
 # Ignore any pycache directories
 **/__pycache__
 .vscode/
+
+scripts/
