+++ Toon.py
@@ -6,7 +6,7 @@ from typing import Optional
 from core.Entity import Entity
 from core.GagGlobals import DEFAULT_GAG_LIMIT
 from core.ToonGlobals import DEFAULT_BEAN_COUNT, DEFAULT_BEAN_LIMIT, DEFAULT_HP
-from Gag import Gags
+from gags.Gag import Gags
 
 
 @dataclass
