+++ core/Factory.py
@@ -1,7 +1,7 @@
 from battle.Battle import BattleCog, ToonAttack
 from Cog import Cog
 from core.Entity import BattleEntity, Entity
-from Gag import Gag, get_gag_min_max_exp
+from gags.Gag import Gag, get_gag_min_max_exp
 from Toon import Toon
 
 from .GagGlobals import TRACK
