+++ tests/battle/test_battle_toon_creation.py
@@ -4,7 +4,7 @@ from core.Entity import BattleEntity, Entity
 from core.GagGlobals import GAG
 from core.ToonGlobals import (ASTRO_HP, ASTRO_NAME, DEFAULT_HP, TRAPA_HP,
                               TRAPA_NAME)
-from Gag import Gag, get_default_gags
+from gags.Gag import Gag, get_default_gags
 from Toon import Inventory, Toon
 
 BATTLE_ID = 0
