+++ tests/fixtures/gag_fixtures.py
@@ -1,8 +1,12 @@
 import pytest
 
-from core.ToonGlobals import (ASTRO_GAG_COUNT, ASTRO_TRACK_EXPS,
-                              TRAPA_GAG_COUNT, TRAPA_TRACK_EXPS)
-from Gag import Gags
+from core.ToonGlobals import (
+    ASTRO_GAG_COUNT,
+    ASTRO_TRACK_EXPS,
+    TRAPA_GAG_COUNT,
+    TRAPA_TRACK_EXPS,
+)
+from gags.Gag import Gags
 
 
 @pytest.fixture
